Aspects of game split into modules
