+++ Blackjack/modules/playing_cards.py
@@ -1,0 +1,24 @@
+import random
+
+def create_shuffled_deck():
+    suits = ["Diamonds", "Hearts", "Spades", "Clubs"]
+    values = ["2", "3", "4", "5", "6", "7", "8", "9", "10", "Jack", "Queen", "King", "Ace"]
+    deck = [f"{value} of {suit}" for suit in suits for value in values]
+    random.shuffle(deck)
+    return deck
+
+scores = {
+    "2": 2,
+    "3": 3,
+    "4": 4,
+    "5": 5,
+    "6": 6,
+    "7": 7,
+    "8": 8,
+    "9": 9,
+    "10": 10,
+    "Jack": 10,
+    "Queen": 10,
+    "King": 10,
+    "Ace": 11
+}
